pyproject.toml
[tool.poetry]
name = "rag-flink"
version = "0.1.0"
description = "RAG Flink Project"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.0"
confluent-kafka = "^2.3.0"
avro-python3 = "^1.10.2"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
pydantic = "^2.6.1"
loguru = "^0.7.2"
langchain = "^0.1.9"
langchain-google-genai = "^0.0.11"
google-generativeai = "^0.4.1"
psutil = "^5.9.8"
PyPDF2 = "^3.0.1"
requests = "^2.31.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 